MAINT: Refactoring interpolators to live in their own files.
