+++ podpac/interpolators.py
@@ -7,4 +7,6 @@ Interpolators Public Module
 
 from podpac.core.interpolation.interpolation import Interpolate
 from podpac.core.interpolation.interpolator import Interpolator
-from podpac.core.interpolation.interpolators import NearestNeighbor, NearestPreview, Rasterio, ScipyGrid, ScipyPoint
+from podpac.core.interpolation.nearest_neighbor import NearestNeighbor, NearestPreview
+from podpac.core.interpolation.rasterio import Rasterio
+from podpac.core.interpolation.scipy import ScipyGrid, ScipyPoint
